添加新接口优化项目结构

+++ model/request/common.go
@@ -2,8 +2,8 @@ package request
 
 // Paging common input parameter structure
 type PageInfo struct {
-	Page     int `json:"page" form:"page"`
-	PageSize int `json:"pageSize" form:"pageSize"`
+	Page     int64 `json:"page" form:"page"`
+	PageSize int64 `json:"pageSize" form:"pageSize"`
 }
 
 // Find by id structure
